Move CorruptedGenerator and support code to its own file
