+++ stellargraph/mapper/__init__.py
@@ -33,3 +33,4 @@ from .graphwave_generator import *
 from .adjacency_generators import *
 from .knowledge_graph import *
 from .padded_graph_generator import *
+from .corrupted import *
